minidump: revert minidump type. dw_text & wic_text: fix bug that text missing when IDWriteRenderer::DrawGlyphRun() is called multiple times by a IDWriteTextLayout::Draw(). dw_text & wic_text: get font properties without creating IDWriteFont when creating IDWriteTextFormat. ft_text: support negative artificial embolden. gdi_text: Removing cursor advancement on y-axis.
--HG--
branch : DirectWrite

+++ gdimm/wic_text.h
@@ -29,7 +29,7 @@ class gdimm_wic_text : public gdimm_text
 	bool _use_gdi_natural;
 	gdimm_wic_dib _wic_bitmap;
 
-	LONG _em_size;
+	FLOAT _em_size;
 	POINT _cursor;
 
 	bool prepare(LPCWSTR lpString, UINT c, text_metrics &metrics, IDWriteFontFace **dw_font_face, DWRITE_GLYPH_RUN &dw_glyph_run);
